Enable webview blur on lost focus
Some webviews maintain internal state that tells them whether they think they are focused or not.

This internal state doesn't get updated when the user launches an editor window.  Instead, existing code relies on other mechanisms to keep this in place, and sometimes this gets out of sync.

This change adds a new method to a webview called "blur()" and invokes it on the active webview, IF that webview thinks it has focus, when a new editor window is opened.